requirements.txt
djangorestframework==3.11.0
Django==2.2
elasticsearch==7.7.1
